refactor: Move script tag position from body to head
the custom script App.js is moved to the <head> of the document. This is to made to take advantage of the attribute defer, that allows the script to be downloaded without stopping the HTML parser.

With the defer attribute, the script will be downloaded in parallel and will be executed after the parsing is completed.

More info at: https://stackoverflow.com/questions/436411/where-should-i-put-script-tags-in-html-markup