fix: Change scripts attributes from async to defer
Scripts using CDN weren't working with async attributes. As this was already mentioned before as a possible problem. Probably happens because async downloads parallel to the HTML parser, but runs as soons as downloads are complete, which makes them run out of order.

With the defer attribute this was corrected, probably because defer still downloads asynchronously, but runs the scripts in order after the HTML was completely parsed.

See more about this here:
https://stackoverflow.com/questions/436411/where-should-i-put-script-tags-in-html-markup