fix: Improve animations
Animation to turn cards was only working for the first card. At the second card, matching or mismatching animation was taking place instead.

This was fixed with delays for the matching and mismatching cards. Also the delay to turn the cards back to the table was increased when cards are mismatched. This is to allow the animations to take place and to allow an animation pause for the player to see the card before it is turned back.