fix: Change time display format on the leaderboard
Change from displaying only seconds to display minutes and seconds in mm:ss format as in the score panel and in the congratulations screen.

Displaying only seconds would look sloppy for game times above one minute as it would show as 60s and above. Now it is showing as 1:00 and so on.